pyproject.toml
[tool.poetry]
name = "A2A"
version = "0.1.0"
description = "Project to look at A2A communication"
authors = ["Benjamin"]

[tool.poetry.dependencies]
python = "^3.13"
pydantic = "^2.11.9"
fastapi = "^0.117.1"
langchain-openai = "^0.3.33"
langgraph = "^0.6.7"
mcp = "^1.15.0"
uvicorn = "^0.34.0"
a2a-sdk = "^0.3.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry-core.masonry.api"
